requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Authlib==1.3.1
python-dotenv==1.0.1
